Rename evm folder to vm, make it as an abstraction general layer
