+++ Cargo.toml
@@ -2,19 +2,19 @@
 members = [
 	"frame/dynamic-fee",
 	"frame/ethereum",
-	"frame/evm",
-	"frame/evm/precompile/simple",
-	"frame/evm/precompile/modexp",
-	"frame/evm/precompile/ed25519",
-	"frame/evm/precompile/bn128",
-	"frame/evm/precompile/blake2",
-	"frame/evm/precompile/dispatch",
+	"frame/vm",
+	"frame/vm/precompile/simple",
+	"frame/vm/precompile/modexp",
+	"frame/vm/precompile/ed25519",
+	"frame/vm/precompile/bn128",
+	"frame/vm/precompile/blake2",
+	"frame/vm/precompile/dispatch",
 	"client/consensus",
 	"client/rpc-core",
 	"client/rpc",
 	"client/db",
 	"primitives/consensus",
-	"primitives/evm",
+	"primitives/vm",
 	"primitives/rpc",
 	"template/node",
 	"template/runtime",
