+++ frame/ethereum/Cargo.toml
@@ -14,11 +14,11 @@ frame-support = { version = "2.0.0-dev", default-features = false, git = "https:
 frame-system = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 pallet-balances = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 pallet-timestamp = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
-pallet-evm = { version = "2.0.0-dev", default-features = false, path = "../evm" }
+pallet-vm = { version = "2.0.0-dev", default-features = false, path = "../vm" }
 sp-runtime = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 sp-std = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 sp-io = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
-fp-evm = { version = "0.8.0", default-features = false, path = "../../primitives/evm" }
+fp-vm = { version = "0.8.0", default-features = false, path = "../../primitives/vm" }
 evm = { version = "0.23.0", features = ["with-codec"], default-features = false }
 ethereum = { version = "0.6", default-features = false, features = ["with-codec"] }
 ethereum-types = { version = "0.10", default-features = false }
@@ -42,10 +42,10 @@ std = [
 	"frame-system/std",
 	"pallet-balances/std",
 	"pallet-timestamp/std",
-	"pallet-evm/std",
+	"pallet-vm/std",
 	"sp-io/std",
 	"sp-std/std",
-	"fp-evm/std",
+	"fp-vm/std",
 	"ethereum/std",
 	"ethereum-types/std",
 	"rlp/std",
