+++ frame/vm/precompile/dispatch/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "pallet-vm-precompile-dispatch"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "DISPATCH precompiles for EVM pallet."
+
+[dependencies]
+sp-core = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+sp-io = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+frame-support = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+pallet-vm = { version = "2.0.0", default-features = false, path = "../.." }
+fp-vm = { version = "0.8.0", default-features = false, path = "../../../../primitives/vm" }
+evm = { version = "0.23.0", default-features = false, features = ["with-codec"] }
+codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"sp-core/std",
+	"sp-io/std",
+	"frame-support/std",
+	"pallet-vm/std",
+	"fp-vm/std",
+	"evm/std",
+	"codec/std",
+]
