+++ frame/vm/precompile/simple/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "pallet-vm-precompile-simple"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "Simple precompiles for EVM pallet."
+
+[dependencies]
+sp-core = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+sp-io = { version = "2.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
+fp-vm = { version = "0.8.0", default-features = false, path = "../../../../primitives/vm" }
+evm = { version = "0.23.0", default-features = false, features = ["with-codec"] }
+ripemd160 = { version = "0.9", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"sp-core/std",
+	"sp-io/std",
+	"fp-vm/std",
+	"evm/std",
+	"ripemd160/std",
+]
