--- primitives/evm/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "fp-evm"
-version = "0.8.0"
-license = "Apache-2.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-description = "Primitive EVM abstractions for Substrate."
-documentation = "https://docs.rs/sp-evm"
-readme = "README.md"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-sp-core = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "frontier", default-features = false }
-sp-std = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "frontier", default-features = false }
-serde = { version = "1.0.101", optional = true, features = ["derive"] }
-codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false }
-evm = { version = "0.23.0", default-features = false, features = ["with-codec"] }
-impl-trait-for-tuples = "0.1"
-
-[features]
-default = ["std"]
-std = [
-	"sp-core/std",
-	"sp-std/std",
-	"serde",
-	"codec/std",
-	"evm/std",
-	"evm/with-serde",
-]
