+++ primitives/rpc/Cargo.toml
@@ -9,7 +9,7 @@ license = "Apache-2.0"
 [dependencies]
 sp-core = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 sp-api = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
-fp-evm = { version = "0.8.0", default-features = false, path = "../../primitives/evm" }
+fp-vm = { version = "0.8.0", default-features = false, path = "../../primitives/vm" }
 ethereum = { version = "0.6", default-features = false, features = ["with-codec"] }
 ethereum-types = { version = "0.10", default-features = false }
 codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false }
@@ -22,7 +22,7 @@ default = ["std"]
 std = [
 	"sp-core/std",
 	"sp-api/std",
-	"fp-evm/std",
+	"fp-vm/std",
 	"ethereum/std",
 	"ethereum-types/std",
 	"codec/std",
