+++ template/runtime/Cargo.toml
@@ -20,8 +20,8 @@ frame-system = { version = "2.0.0-dev", default-features = false, package = "fra
 frame-system-rpc-runtime-api = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 
 pallet-ethereum = { version = "0.1.0", default-features = false, path = "../../frame/ethereum" }
-pallet-evm = { version = "2.0.0-dev", default-features = false, path = "../../frame/evm" }
-pallet-evm-precompile-simple = { version = "2.0.0-dev", default-features = false, path = "../../frame/evm/precompile/simple" }
+pallet-vm = { version = "2.0.0-dev", default-features = false, path = "../../frame/vm" }
+pallet-vm-precompile-simple = { version = "2.0.0-dev", default-features = false, path = "../../frame/vm/precompile/simple" }
 pallet-aura = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 pallet-balances = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
 pallet-grandpa = { version = "2.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "frontier" }
@@ -62,8 +62,8 @@ std = [
 	"fp-rpc/std",
 
 	"pallet-ethereum/std",
-	"pallet-evm/std",
-	"pallet-evm-precompile-simple/std",
+	"pallet-vm/std",
+	"pallet-vm-precompile-simple/std",
 	"pallet-aura/std",
 	"pallet-balances/std",
 	"pallet-grandpa/std",
