+++ ts-tests/frontier-test-specs/simple-specs.json
@@ -25,7 +25,7 @@
         "key": "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY"
       },
       "palletEthereum": {},
-      "palletEvm": {
+      "palletVm": {
         "accounts": {
           "0x6be02d1d3665660d22ff9624b7be0551ee1ac91b": {
             "nonce": "0x0",
